+++ examples/parquet/py/avro_in.py
@@ -18,60 +18,11 @@
 #
 # END_COPYRIGHT
 
-import sys
-from cStringIO import StringIO
 from collections import Counter
 
-import pydoop
 import pydoop.mapreduce.api as api
 import pydoop.mapreduce.pipes as pp
-
-from avro.io import DatumReader, BinaryDecoder
-import avro
-
-
-AVRO_INPUT = pydoop.PROPERTIES['AVRO_INPUT']
-AVRO_VALUE_INPUT_SCHEMA = pydoop.PROPERTIES['AVRO_VALUE_INPUT_SCHEMA']
-
-
-def get_schema(jc):
-    """
-    Get schema from JSON string.
-    """
-    schema_str = jc.get(AVRO_VALUE_INPUT_SCHEMA)
-    return avro.schema.parse(schema_str)
-
-
-def get_schema_alt(jc):
-    """
-    Get schema from parsed JSON (alternate method for doc purposes).
-    """
-    schema_json = jc.get_json(AVRO_VALUE_INPUT_SCHEMA)
-    return avro.schema.make_avsc_object(schema_json)
-
-
-class AvroContext(pp.TaskContext):
-
-    datum_reader = None  # FIXME not strictly necessary
-
-    def set_job_conf(self, vals):
-        super(AvroContext, self).set_job_conf(vals)
-        jc = self.get_job_conf()
-        # This method is called both in map and reduce tasks.  Since AVRO_INPUT
-        # and AVRO_VALUE_INPUT_SCHEMA are set by PydoopAvroBridgeReader,
-        # however, they will only be present in the map task's conf.
-        if AVRO_INPUT in jc:
-            assert jc.get(AVRO_INPUT).upper() == 'V'
-            schema = get_schema(jc)
-            assert get_schema_alt(jc).to_json() == schema.to_json()
-            self.datum_reader = DatumReader(schema)
-
-    def get_input_value(self):
-        # FIXME reuse, reuse, reuse
-        sys.stderr.write('value: %r\n' % self._value)
-        f = StringIO(self._value)
-        dec = BinaryDecoder(f)
-        return self.datum_reader.read(dec)
+from pydoop.avrolib import AvroContext
 
 
 class ColorPick(api.Mapper):
@@ -79,7 +30,6 @@ class ColorPick(api.Mapper):
     def map(self, ctx):
         user = ctx.value
         color = user['favorite_color']
-        sys.stderr.write('user: %r\n' % (user,))
         if color is not None:
             ctx.emit(user['office'], Counter({color: 1}))
